requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
astunparse==1.6.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
chromedrivermanager==0.0.1
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
Django==5.1.2
django-model-utils==5.0.0
djangorestframework==3.15.2
fastapi==0.115.2
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.9.0
gast==0.6.0
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-pasta==0.2.0
googleapis-common-protos==1.65.0
grpcio==1.66.1
h11==0.14.0
httplib2==0.22.0
httptools==0.6.1
humanfriendly==10.0
idna==3.7
Jinja2==3.1.4
keras==3.5.0
kiwisolver==1.4.7
libclang==18.1.1
MarkupSafe==3.0.1
matplotlib==3.9.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
networkx==3.4
numpy==2.1.2
onnx==1.17.0
onnxruntime==1.19.2
onnxslim==0.1.34
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.0.0
py-cpuinfo==9.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.1.4
pypiwin32==223
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
roboflow==1.1.44
rsa==4.9
scipy==1.14.1
seaborn==0.13.2
selenium==4.25.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.39.2
sympy==1.13.3
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow-intel==2.17.0
termcolor==2.4.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.10
ultralytics-thop==2.0.9
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.31.1
watchfiles==0.24.0
webdriver-manager==4.0.2
websockets==13.1
wheel==0.44.0
wsproto==1.2.0
